Log improvement
